# public/documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BileMo APi",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://127.0.0.1:8000/api/v1",
            "description": "BileMo API"
        }
    ],
    "paths": {
        "/api/v1/clients ": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Add new client",
                "operationId": "App\\Controller\\ClientController::register",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Client object that needs to be added in database",
                        "example": {"firstname": "Daniel", "lastname": "Grondin", "email": "mailclient@gmail.com", "street_number": 26, "street": "barriere de saumur", "cp": 85200, "city": "Fontenay", "phone_number": "0606060606" },
                        "required": true
                    }

                ],
                "responses": {
                    "201": {
                        "description": "client created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}